tools/requirements.txt
jsonschema>=4.17.0
jcs>=0.2.1
PyYAML>=6.0.1
